VendorMicro getting ready for auth as well now.
